Final Refactoring
