+++ src/mtf/dm/cms/hhs/gov/tests/demoUI/pages/adminUserManagementPage.ts
@@ -1,5 +1,5 @@
 import { Page } from '@playwright/test';
-import { UIActionUtilities } from '../../../utilities/UIActionUtilities';
+import { UIActionUtilities } from '../../../commonUtilities/UIActionUtilities';
 import { TestScenarioContext } from '../context/TestScenarioContext';
